feat(practitioner): brand deploy site, root auth redirects, and /mfa enrollment
