+++ .env.example
@@ -51,7 +51,10 @@ NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY=sb_publishable_your_key
 
 # apps/web: canonical site URL for SEO (metadataBase, sitemap, Open Graph). Production: https://abstrack.org
 # apps/mobile: use EXPO_PUBLIC_USER_WEB_ORIGIN with the same origin (Universal Links / App Links).
-NEXT_PUBLIC_USER_WEB_ORIGIN=https://abstrack.org
+NEXT_PUBLIC_USER_WEB_ORIGIN=http://localhost:3000
+
+# apps/practitioner: canonical origin for metadataBase / Open Graph. Production: https://practitioner.abstrack.org
+# NEXT_PUBLIC_PRACTITIONER_WEB_ORIGIN=http://localhost:3000
 
 # apps/practitioner: set `NEXT_PUBLIC_APP_NAME` so the TOTP issuer in authenticator apps matches
 # your product name per environment. If unset, the issuer falls back to "ABStrack".
@@ -72,7 +75,7 @@ NEXT_PUBLIC_APP_NAME=ABStrack
 EXPO_PUBLIC_SUPABASE_URL=https://YOUR_PROJECT_REF.supabase.co
 EXPO_PUBLIC_SUPABASE_PUBLISHABLE_KEY=sb_publishable_your_key
 # Same origin as NEXT_PUBLIC_USER_WEB_ORIGIN on user web (caretaker invite / App Links).
-EXPO_PUBLIC_USER_WEB_ORIGIN=https://abstrack.org
+EXPO_PUBLIC_USER_WEB_ORIGIN=http://localhost:3000
 
 # PowerSync Service HTTP/WebSocket endpoint for mobile offline replication (configure instance + upload `packages/powersync/sync-rules.yaml`).
 # Copy the current instance URL from the PowerSync Dashboard — it may change if the Cloud project was disabled and re-enabled (see docs/SUPABASE_CLOUD_DEVELOPER.md).
