+++ apps/practitioner/src/components/practitioner-patient-routes-gate.tsx
@@ -220,7 +220,7 @@ export function PractitionerPatientRoutesGate({
           <div className="mt-6 flex flex-wrap gap-3">
             {mfaBlockReason === 'enrollment' ? (
               <Link
-                href="/"
+                href="/mfa"
                 className="inline-flex min-h-11 items-center justify-center rounded-md bg-app-primary-solid px-4 py-2 text-sm font-medium text-app-on-primary-solid transition hover:brightness-105 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-app-ring focus-visible:ring-offset-2 focus-visible:ring-offset-app-bg"
               >
                 Set up two-factor authentication
