+++ apps/practitioner/src/lib/practitioner-nav-items.ts
@@ -2,13 +2,10 @@ import type { AppSideNavItem } from '@abstrack/ui-web';
 
 /**
  * Primary practitioner app routes for the side navigation.
+ * Brand / logo links to `/` (proxy redirects signed-in users to `/patients`, signed-out to `/login`).
+ * TOTP setup is under `/mfa` and Settings.
  */
 export const PRACTITIONER_APP_NAV_ITEMS: AppSideNavItem[] = [
-  {
-    href: '/',
-    label: 'Home',
-    match: (path) => path === '/',
-  },
   {
     href: '/patients',
     label: 'Patients',
